Move Auth Packets to own package
