+++ core/src/main/java/de/timuuuu/moneymaker/moneychat/MoneyChatSession.java
@@ -13,10 +13,10 @@ import de.timuuuu.moneymaker.moneychat.protocol.PacketHandler;
 import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketAddonStatistics;
 import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketClearChat;
 import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketDisconnect;
-import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketEncryptionRequest;
-import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketEncryptionResponse;
-import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketHelloPong;
-import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketLoginComplete;
+import de.timuuuu.moneymaker.moneychat.protocol.packets.auth.PacketEncryptionRequest;
+import de.timuuuu.moneymaker.moneychat.protocol.packets.auth.PacketEncryptionResponse;
+import de.timuuuu.moneymaker.moneychat.protocol.packets.auth.PacketHelloPong;
+import de.timuuuu.moneymaker.moneychat.protocol.packets.auth.PacketLoginComplete;
 import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketMessage;
 import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketMessageDelete;
 import de.timuuuu.moneymaker.moneychat.protocol.packets.PacketPing;
