+++ core/src/main/java/de/timuuuu/moneymaker/moneychat/protocol/packets/auth/PacketLoginComplete.java
@@ -1,4 +1,4 @@
-package de.timuuuu.moneymaker.moneychat.protocol.packets;
+package de.timuuuu.moneymaker.moneychat.protocol.packets.auth;
 
 import de.timuuuu.moneymaker.moneychat.protocol.Packet;
 import de.timuuuu.moneymaker.moneychat.protocol.PacketBuffer;
